Changed dropbox permissions
The syncany dropbox plugin previously requested read/write access to all
files in a dropbox. This made it possible for foreigners to read a
user's whole dropbox once he has access on a shared dropbox repository.
This is now changed to be more restrictive: from this point syncany can
only access files which were created using a shared accessToken.