Make URI creating more predictable
Created an UriBuilder to create Uris pointing to a file within a
Dropbox account. Dropbox does  not like URIs beginning with '//',
although, that seems to be ok with Uri.Especially took care that it is
very robust since it is taking user input.
Also added a test to verify the generated Uri.

fixes syncany/syncany#417